+++ lib/src/option.dart
@@ -8,54 +8,94 @@
 /// get to it. This function isn't exported to the public API of the package.
 Option newOption(
     String name,
-    String abbreviation,
+    String abbr,
     String help,
     String valueHelp,
     Iterable<String> allowed,
     Map<String, String> allowedHelp,
-    defaultValue,
+    defaultsTo,
     Function callback,
     OptionType type,
     {bool negatable,
     bool splitCommas,
     bool hide: false}) {
-  return new Option._(name, abbreviation, help, valueHelp, allowed, allowedHelp,
-      defaultValue, callback, type,
+  return new Option._(name, abbr, help, valueHelp, allowed, allowedHelp,
+      defaultsTo, callback, type,
       negatable: negatable, splitCommas: splitCommas, hide: hide);
 }
 
-/// A command-line option. Includes both flags and options which take a value.
+/// A command-line option.
+///
+/// This represents both boolean flags and options which take a value.
 class Option {
+  /// The name of the option that the user passes as an argument.
   final String name;
-  final String abbreviation;
-  final List<String> allowed;
-  final defaultValue;
-  final Function callback;
+
+  /// A single-character string that can be used as a shorthand for this option.
+  ///
+  /// For example, `abbr: "a"` will allow the user to pass `-a value` or
+  /// `-avalue`.
+  final String abbr;
+
+  @Deprecated("Use abbr instead.")
+  String get abbreviation => abbr;
+
+  /// A description of this option.
   final String help;
+
+  /// A name for the value this option takes.
   final String valueHelp;
+
+  /// A list of valid values for this option.
+  final List<String> allowed;
+
+  /// A map from values in [allowed] to documentation for those values.
   final Map<String, String> allowedHelp;
-  final OptionType type;
+
+  /// The value this option will have if the user doesn't explicitly pass it in
+  final defaultsTo;
+
+  @Deprecated("Use defaultsTo instead.")
+  get defaultValue => defaultsTo;
+
+  /// Whether this flag's value can be set to `false`.
+  ///
+  /// For example, if [name] is `flag`, the user can pass `--no-flag` to set its
+  /// value to `false`.
+  ///
+  /// This is `null` unless [type] is [OptionType.flag].
   final bool negatable;
+
+  /// The callback to invoke with the option's value when the option is parsed.
+  final Function callback;
+
+  /// Whether this is a flag, a single value option, or a multi-value option.
+  final OptionType type;
+
+  /// Whether multiple values may be passed by writing `--option a,b` in
+  /// addition to `--option a --option b`.
   final bool splitCommas;
+
+  /// Whether this option should be hidden from usage documentation.
   final bool hide;
 
   /// Whether the option is boolean-valued flag.
-  bool get isFlag => type == OptionType.FLAG;
+  bool get isFlag => type == OptionType.flag;
 
   /// Whether the option takes a single value.
-  bool get isSingle => type == OptionType.SINGLE;
+  bool get isSingle => type == OptionType.single;
 
   /// Whether the option allows multiple values.
-  bool get isMultiple => type == OptionType.MULTIPLE;
+  bool get isMultiple => type == OptionType.multiple;
 
   Option._(
       this.name,
-      this.abbreviation,
+      this.abbr,
       this.help,
       this.valueHelp,
       Iterable<String> allowed,
       Map<String, String> allowedHelp,
-      this.defaultValue,
+      this.defaultsTo,
       this.callback,
       OptionType type,
       {this.negatable,
@@ -68,7 +108,7 @@ class Option {
         // If the user doesn't specify [splitCommas], it defaults to true for
         // multiple options.
         this.splitCommas =
-            splitCommas == null ? type == OptionType.MULTIPLE : splitCommas {
+            splitCommas == null ? type == OptionType.multiple : splitCommas {
     if (name.isEmpty) {
       throw new ArgumentError('Name cannot be empty.');
     } else if (name.startsWith('-')) {
@@ -80,14 +120,14 @@ class Option {
       throw new ArgumentError('Name "$name" contains invalid characters.');
     }
 
-    if (abbreviation != null) {
-      if (abbreviation.length != 1) {
+    if (abbr != null) {
+      if (abbr.length != 1) {
         throw new ArgumentError('Abbreviation must be null or have length 1.');
-      } else if (abbreviation == '-') {
+      } else if (abbr == '-') {
         throw new ArgumentError('Abbreviation cannot be "-".');
       }
 
-      if (_invalidChars.hasMatch(abbreviation)) {
+      if (_invalidChars.hasMatch(abbr)) {
         throw new ArgumentError('Abbreviation is an invalid character.');
       }
     }
@@ -96,15 +136,13 @@ class Option {
   /// Returns [value] if non-`null`, otherwise returns the default value for
   /// this option.
   ///
-  /// For single-valued options, it will be [defaultValue] if set or `null`
+  /// For single-valued options, it will be [defaultsTo] if set or `null`
   /// otherwise. For multiple-valued options, it will be an empty list or a
-  /// list containing [defaultValue] if set.
+  /// list containing [defaultsTo] if set.
   dynamic getOrDefault(value) {
     if (value != null) return value;
-
-    if (!isMultiple) return defaultValue;
-    if (defaultValue != null) return [defaultValue];
-    return [];
+    if (isMultiple) return defaultsTo ?? <String>[];
+    return defaultsTo;
   }
 
   static final _invalidChars = new RegExp(r'''[ \t\r\n"'\\/]''');
@@ -115,7 +153,10 @@ class OptionType {
   /// An option that can only be `true` or `false`.
   ///
   /// The presence of the option name itself in the argument list means `true`.
-  static const FLAG = const OptionType._("OptionType.FLAG");
+  static const flag = const OptionType._("OptionType.flag");
+
+  @Deprecated("Use OptionType.flag instead.")
+  static const FLAG = flag;
 
   /// An option that takes a single value.
   ///
@@ -126,7 +167,10 @@ class OptionType {
   ///     --mode=debug
   ///
   /// If the option is passed more than once, the last one wins.
-  static const SINGLE = const OptionType._("OptionType.SINGLE");
+  static const single = const OptionType._("OptionType.single");
+
+  @Deprecated("Use OptionType.single instead.")
+  static const SINGLE = single;
 
   /// An option that allows multiple values.
   ///
@@ -136,7 +180,10 @@ class OptionType {
   ///
   /// In the parsed `ArgResults`, a multiple-valued option will always return
   /// a list, even if one or no values were passed.
-  static const MULTIPLE = const OptionType._("OptionType.MULTIPLE");
+  static const multiple = const OptionType._("OptionType.multiple");
+
+  @Deprecated("Use OptionType.multiple instead.")
+  static const MULTIPLE = multiple;
 
   final String name;