+++ test/command_test.dart
@@ -91,10 +91,7 @@ Run "test help" to see global options."""));
 
   test("usageException splits up the message and usage", () {
     expect(
-        () => foo.usageException("message"),
-        throwsUsageException(
-            "message",
-            """
+        () => foo.usageException("message"), throwsUsageException("message", """
 Usage: test foo [arguments]
 -h, --help    Print this usage information.